Move Flash-specific logic to `to-maps.js`
