+++ bin/cli.js
@@ -10,7 +10,6 @@ const rimraf = require('rimraf');
 
 const convertFromMaps = require('../src/from-maps.js');
 const convertToMaps = require('../src/to-maps.js');
-const convertToFlash = require('../src/to-flash.js');
 const generateBounds = require('../src/generate-bounds.js');
 const info = require('../package.json');
 
@@ -32,6 +31,7 @@ const main = function() {
 		console.log('\nUsage:\n');
 		console.log(`\t${info.name} --from-maps=./Automap --output-dir=./data`);
 		console.log(`\t${info.name} --from-data=./data --output-dir=./Automap --no-markers`);
+		console.log(`\t${info.name} --from-data=./data-without-markers --output-file=./flash/maps-without-markers.exp --flash --no-markers`);
 		process.exit(1);
 	}
 
@@ -75,23 +75,21 @@ const main = function() {
 			argv['from-data'] = 'data';
 		}
 		const dataDirectory = path.resolve(argv['from-data']);
-
-		if (argv['flash'] === true) {
-			if (!argv['output-file'] || argv['output-file'] === true) {
-				console.log('`--output-file` path not specified.');
-				return process.exit(1);
-			}
-			convertToFlash(dataDirectory, argv['output-file'], !excludeMarkers);
+		if (argv['flash-export-file'] === true) {
+			console.log('`--flash-export-file` path not specified.');
+			return process.exit(1);
+		}
+		if (argv['flash-export-file']) {
+			convertToMaps(dataDirectory, argv['flash-export-file'], !excludeMarkers, true);
 			return;
 		}
-
 		if (!argv['output-dir'] || argv['output-dir'] === true) {
 			console.log('`--output-dir` path not specified. Using the default, i.e. `Automap-new`.');
 			argv['output-dir'] = 'Automap-new';
 		}
 		const mapsDirectory = path.resolve(String(argv['output-dir']));
 		emptyDirectory(mapsDirectory).then(function() {
-			convertToMaps(dataDirectory, mapsDirectory, !excludeMarkers);
+			convertToMaps(dataDirectory, mapsDirectory, !excludeMarkers, false);
 		});
 		return;
 	}
