+++ src/array-to-flash-markers.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const windows1252 = require('windows-1252');
+
 const icons = require('./icons.js');
 
 const arrayToFlashMarkers = function(array, floorID) {
@@ -8,10 +10,7 @@ const arrayToFlashMarkers = function(array, floorID) {
 		// Note: the property order is important here. This order matches the
 		// `*.exp` file format exactly.
 		result.push({
-			// Note that `*.exp` files created by the C++ Tibia client replace
-			// any non-ASCII symbols with `?`. We could do that here as well, but
-			// since the migration tool handles such symbols just fine, why bother?
-			'description': marker.description,
+			'description': windows1252.encode(marker.description),
 			'type': icons.byName[marker.icon],
 			'x': marker.xTile * 256 + marker.xPosition,
 			'y': marker.yTile * 256 + marker.yPosition,
