+++ src/to-maps.js
@@ -9,9 +9,12 @@ const padLeft = require('lodash.padleft');
 const handleSequence = require('./handle-sequence.js');
 const writeJSON = require('./write-json.js');
 
+const arrayToFlashMarkers = require('./array-to-flash-markers.js');
+const arrayToMarkerBuffer = require('./array-to-marker.js');
+const idToXyz = require('./id-to-xyz.js');
 const pixelDataToMapBuffer = require('./pixel-data-to-map.js');
 const pixelDataToPathBuffer = require('./pixel-data-to-path.js');
-const arrayToMarkerBuffer = require('./array-to-marker.js');
+const transposeBuffer = require('./transpose-buffer.js');
 
 const GLOBALS = {};
 
@@ -82,20 +85,21 @@ const createBinaryMarkers = function(floorID) {
 			const markers = data[id];
 			const markerBuffer = arrayToMarkerBuffer(markers);
 			addResult(id, 'markerBuffer', markerBuffer);
+			const flashMarkers = arrayToFlashMarkers(markers, floorID);
+			addResult(id, 'flashMarkers', flashMarkers);
 		});
 		resolve();
 	});
 };
 
-const convertToMaps = function(dataDirectory, mapsDirectory, includeMarkers) {
+const convertToMaps = function(dataDirectory, outputPath, includeMarkers, isFlash) {
 	if (!dataDirectory) {
 		dataDirectory = 'data';
 	}
-	if (!mapsDirectory) {
-		mapsDirectory = 'Automap-new';
+	if (!isFlash && !outputPath) {
+		outputPath = 'Automap-new';
 	}
 	GLOBALS.dataDirectory = dataDirectory;
-	GLOBALS.mapsDirectory = mapsDirectory;
 	const bounds = JSON.parse(fs.readFileSync(`${dataDirectory}/bounds.json`));
 	GLOBALS.bounds = bounds;
 	GLOBALS.canvas = new Canvas(bounds.width, bounds.height);
@@ -109,6 +113,25 @@ const convertToMaps = function(dataDirectory, mapsDirectory, includeMarkers) {
 		}
 	}).then(function() {
 		const noMarkersBuffer = new Buffer([0x00, 0x00, 0x00, 0x00]);
+		if (isFlash) {
+			const lines = Object.keys(RESULTS).map(function(id) {
+				const coordinates = idToXyz(id);
+				const data = RESULTS[id];
+				const entry = {
+					'colordata': transposeBuffer(data.mapBuffer).toString('base64'),
+					'mapmarkers': data.flashMarkers || [],
+					'waypoints': transposeBuffer(data.pathBuffer).toString('base64'),
+					'x': coordinates.x * 256,
+					'y': coordinates.y * 256,
+					'z': coordinates.z
+				};
+				return JSON.stringify(entry);
+			});
+			const contents = lines.join('\r\n') + '\r\n';
+			fs.writeFileSync(outputPath, contents, 'binary');
+			console.log(`${outputPath} created successfully.`);
+			return;
+		}
 		Object.keys(RESULTS).forEach(function(id) {
 			const data = RESULTS[id];
 			const buffer = Buffer.concat([
@@ -116,7 +139,7 @@ const convertToMaps = function(dataDirectory, mapsDirectory, includeMarkers) {
 				data.pathBuffer,
 				includeMarkers ? data.markerBuffer || noMarkersBuffer : noMarkersBuffer
 			]);
-			const fileName = `${mapsDirectory}/${id}.map`;
+			const fileName = `${outputPath}/${id}.map`;
 			const writeStream = fs.createWriteStream(fileName);
 			writeStream.write(buffer);
 			writeStream.end();
